Extract http_utils from run_batch_standardization
Second slice of the Wave C decomposition. Pulls the URL retry/backoff
client (_read_url_response, _read_url_bytes), the landing-page HTML
parser (_LandingPageParser), the Cloudflare/login challenge detector
(_looks_like_challenge_response, _build_web_dataset_access_error), the
SourceAccessError type, and all download-payload helpers
(_normalize_remote_filename, _build_unique_destination,
_extract_content_disposition_filename, _infer_extension_from_content_type,
_is_supported_download_response, _looks_like_supported_download_url,
_decode_http_text, _iter_content_urls, _extract_html_download_urls,
_write_download_payload) into a new self-contained scripts/http_utils
module along with the constants they consult (HTTP_USER_AGENT,
NETWORK_TIMEOUTS, WEB_LOGIN_MARKERS, WEB_NO_DATA_MARKERS{,_ASCII},
WEB_CHALLENGE_HOSTS, WEB_CHALLENGE_TITLE_MARKERS,
WEB_CHALLENGE_BODY_MARKERS).

The new module imports only from stdlib + scripts.archive_utils (for
the suffix sets used by _looks_like_supported_download_url and
_is_supported_download_response). Like archive_utils it can be unit
tested without dragging in pandas / transformers.

run_batch_standardization re-exports every previously-public name so
the Wave B test suite continues to import _read_url_response,
_looks_like_challenge_response, _build_web_dataset_access_error, etc.
directly from scripts.run_batch_standardization. Stdlib imports that
the orchestrator no longer needs (socket, html.unescape, html.parser,
urljoin, unquote) are dropped from its top of file.

Net diff in the orchestrator: -385 / +31 lines. Test result: 281 passed,
1 skipped (unchanged).
